+++ source/core/calibration/Calibration.cpp
@@ -1,4 +1,4 @@
-#include <core/calibration/calibration.h>
+#include <core/calibration/Calibration.h>
 #include <core/calibration/coef_function_factory.h>
 #include <core/util/string_extensions.h>
 #include <core/util/time_extensions.h>
