+++ tests/core/calibration/CalibrationTest.cpp
@@ -1,5 +1,5 @@
 #include "gtest_color_print.h"
-#include <core/calibration/calibration.h>
+#include <core/calibration/Calibration.h>
 #include <core/calibration/coef_function_factory.h>
 #include <core/calibration/Polynomial.h>
